+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # ItemSense Node Library Changelog
 
+## 1.1.0
+Adds reader health behaviour.
+- All health endpoints added: ( `events`, `configureQueue`, `readers`, and `reader` )
 
 ## 1.0.9a
 Now accepts `authToken` upon instantiation, as documented.